created route and server.js, MVP met
